pr01/requirements.txt
numpy==1.18.4
Pillow==7.1.2
progress==1.5
